refactor

+++ CodingTracker/Controllers/CodingSessionController.cs
@@ -6,36 +6,47 @@ namespace CodingTracker.Controllers;
 
 internal class CodingSessionController
 {
-    private readonly Database db = new Database();
-    public void RecordSession(bool update = false)
+    public CodingSession? CreateSession()
     {
-        DateTime startTime;
-        DateTime endTime;
-        int id = 0;
-
-        if (update)
-        {
-            DisplaySessions();
-            id = AnsiConsole.Ask<int>("Enter the [green]ID[/] of the session you want to update:");
-        }
-
-        startTime = InputHelper.GetDateInput("start");
-        endTime = InputHelper.GetDateInput("end");
+        DateTime startTime = InputHelper.GetDateInput("start");
+        DateTime endTime = InputHelper.GetDateInput("end");
 
         if (startTime > endTime)
         {
-            AnsiConsole.MarkupLine("[red]End time cannot be before start time.[/]");
-            return;
+            return null;
         }
-
-        var session = new CodingSession
+        else
         {
-            StartTime = startTime.ToString("dd.MM.yyyy HH:mm"),
-            EndTime = endTime.ToString("dd.MM.yyyy HH:mm"),
-            Duration = InputHelper.CalculateDuration(startTime, endTime)
-        };
+            return new CodingSession
+            {
+                StartTime = startTime.ToString("dd.MM.yyyy HH:mm"),
+                EndTime = endTime.ToString("dd.MM.yyyy HH:mm"),
+                Duration = InputHelper.CalculateDuration(startTime, endTime)
+            };
+        }
+    }
+    public void RecordSession(DatabaseController db)
+    {
+        var session = CreateSession();
+        if (session != null)
+        {
+            db.AddSession(session);
+            AnsiConsole.MarkupLine("[green]Session added successfully[/]");
+        }
+        else
+        {
+            AnsiConsole.MarkupLine("[red]End time cannot be earlier than start time.[/]");
+        }
+    }
 
-        if (update)
+    public void UpdateSession(DatabaseController db)
+    {
+        DisplaySessions(db);
+        int id = AnsiConsole.Ask<int>("\nEnter the [green]ID[/] of the session you want to update or enter 0 to cancel:");
+        if (id == 0) 
+            return;
+        var session = CreateSession();
+        if (session != null)
         {
             if (db.UpdateSession(id, session))
             {
@@ -48,45 +59,33 @@ internal class CodingSessionController
         }
         else
         {
-            db.AddSession(session);
-            AnsiConsole.MarkupLine("[green]Session recorded successfully.[/]");
+            AnsiConsole.MarkupLine("[red]End time cannot be earlier than start time.[/]");
         }
     }
 
-    public void DisplaySessions()
+    public void DisplaySessions(DatabaseController db)
     {
         List<CodingSession> sessions = db.GetSessions();
-
         if (sessions.Count == 0)
         {
             AnsiConsole.MarkupLine("[red]No sessions recorded yet.[/]");
             return;
         }
 
-        var table = new Table();
-        table.Border = TableBorder.Rounded;
-        table.AddColumn("[green]Session ID[/]");
-        table.AddColumn("[green]Start Time[/]");
-        table.AddColumn("[green]End Time[/]");
-        table.AddColumn("[green]Duration (hh:mm)[/]");
-
-        foreach (var session in sessions)
-        {
-            table.AddRow(
-               session.Id.ToString(),
-               session.StartTime,
-               session.EndTime,
-               session.Duration
-               );
-        }
-
+        TimeSpan totalTime = CalculateTotalTime(sessions);
+        TimeSpan averageTime = new(totalTime.Ticks / sessions.Count);
+        var table = InputHelper.GenerateTable(sessions);
         AnsiConsole.Write(table);
+        AnsiConsole.MarkupLine($"Total time spent coding: [green]{totalTime.Hours} hours and {totalTime.Minutes} minutes[/]");
+        AnsiConsole.MarkupLine($"Average time spent coding: [green]{averageTime.Hours} hours and {averageTime.Minutes} minutes[/]");
     }
 
-    public void DeleteSession()
+    public void DeleteSession(DatabaseController db)
     {
-        DisplaySessions();
-        int id = AnsiConsole.Ask<int>("Enter the [green]ID[/] of the session you want to delete:");
+        DisplaySessions(db);
+        int id = AnsiConsole.Ask<int>("\nEnter the [green]ID[/] of the session you want to delete or enter 0 to cancel:");
+        if (id == 0) 
+            return;
         if (db.DeleteSession(id))
         {
             AnsiConsole.MarkupLine("[green]Session deleted successfully.[/]");
@@ -95,5 +94,15 @@ internal class CodingSessionController
         {
             AnsiConsole.MarkupLine("[red]Session not found.[/]");
         }
+    }
+
+    private TimeSpan CalculateTotalTime(List<CodingSession> sessions)
+    {
+        TimeSpan totalTime = new TimeSpan();
+        foreach (var session in sessions)
+        {
+            totalTime += TimeSpan.Parse(session.Duration);
+        }
+        return totalTime;
     }
 }
